Create directories for style, redux and components
