+++ src/App.js
@@ -1,0 +1,9 @@
+import Header from './components/Header';
+
+function App() {
+  return (
+    <Header />
+  );
+}
+
+export default App;
